requirements.txt
# pip install requirements.txt

pyyaml
requests